rename pluginCard => packageCard
